requirements/default.txt
requests>=2.7.0,<3.0
requests_toolbelt
six>=1.9.0,<2.0
python-dateutil>=2.4.2,<3.0
pytz
